requirements.txt
torch
torchvision
numpy
prettytable
Pillow